functions/ingest/requirements.txt
requests==2.31.0
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.16.2